% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{complete_dates}
\alias{complete_dates}
\title{Complete date values}
\usage{
complete_dates(values)
}
\arguments{
\item{values}{A vector of string.}
}
\value{
A vector of string.
}
\description{
Complete date values
}
\keyword{internal}
