% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{complete_times}
\alias{complete_times}
\title{Complete time values}
\usage{
complete_times(values)
}
\arguments{
\item{values}{A vector of string.}
}
\value{
A vector of string.
}
\description{
Complete time values
}
\keyword{internal}
