% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{configure_quarter_level}
\alias{configure_quarter_level}
\alias{configure_quarter_level.when}
\title{Configure quarter level}
\usage{
configure_quarter_level(td, include_all, exclude_all, quarter, year_quarter)

\method{configure_quarter_level}{when}(
  td,
  include_all = FALSE,
  exclude_all = FALSE,
  quarter = NULL,
  year_quarter = NULL
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{include_all}{A boolean, include all fields of the level.}

\item{exclude_all}{A boolean, exclude all fields of the level.}

\item{quarter}{A boolean, include the quarter field.}

\item{year_quarter}{A boolean, include the quarter field.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a date type, using this function we can select
the quarter level and its attributes to include in it: quarter number and the
combination of the year with it.
}
\details{
The \code{include_all} and \code{exclude_all} parameters allow us to include or exclude
all attributes, and then specifically exclude or include the ones we need.
}
\examples{

td <- when() |>
  configure_quarter_level(quarter = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_date_levels}()},
\code{\link{configure_day_level}()},
\code{\link{configure_dimension}()},
\code{\link{configure_month_level}()},
\code{\link{configure_semester_level}()},
\code{\link{configure_time_level}()},
\code{\link{configure_week_level}()},
\code{\link{configure_year_level}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
