% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{configure_semester_level}
\alias{configure_semester_level}
\alias{configure_semester_level.when}
\title{Configure semester level}
\usage{
configure_semester_level(td, include_all, exclude_all, semester, year_semester)

\method{configure_semester_level}{when}(
  td,
  include_all = FALSE,
  exclude_all = FALSE,
  semester = NULL,
  year_semester = NULL
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{include_all}{A boolean, include all fields of the level.}

\item{exclude_all}{A boolean, exclude all fields of the level.}

\item{semester}{A boolean, include the semester field.}

\item{year_semester}{A boolean, include the semester field.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a date type, using this function we can select
the semester level and its attributes to include in it: semester number and the
combination of the year with it.
}
\details{
The \code{include_all} and \code{exclude_all} parameters allow us to include or exclude
all attributes, and then specifically exclude or include the ones we need.
}
\examples{

td <- when() |>
  configure_semester_level(semester = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_date_levels}()},
\code{\link{configure_day_level}()},
\code{\link{configure_dimension}()},
\code{\link{configure_month_level}()},
\code{\link{configure_quarter_level}()},
\code{\link{configure_time_level}()},
\code{\link{configure_week_level}()},
\code{\link{configure_year_level}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
