% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{configure_type}
\alias{configure_type}
\alias{configure_type.when}
\title{Configure type}
\usage{
configure_type(td, type, locale, week_starts_monday)

\method{configure_type}{when}(
  td,
  type = NULL,
  locale = Sys.getlocale("LC_TIME"),
  week_starts_monday = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{type}{A string, type of calendar (NULL, 'iso', 'epi' or 'time').}

\item{locale}{A locale, to use for day and month names.}

\item{week_starts_monday}{A boolean.}
}
\value{
A \code{when} object.
}
\description{
Configure type.
}
\examples{

td <- when() |>
  configure_type()

}
\seealso{
Other time definition: 
\code{\link{configure_instances}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{time_seconds}},
\code{\link{when}()}
}
\concept{time definition}
