% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{configure_year_level}
\alias{configure_year_level}
\alias{configure_year_level.when}
\title{Include year level}
\usage{
configure_year_level(td, include_all, exclude_all, year, decade)

\method{configure_year_level}{when}(
  td,
  include_all = FALSE,
  exclude_all = FALSE,
  year = NULL,
  decade = NULL
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{include_all}{A boolean, include all fields of the level.}

\item{exclude_all}{A boolean, exclude all fields of the level.}

\item{year}{A boolean, include the year field.}

\item{decade}{A boolean, include the decade field.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a date type, using this function we can select
the year level and its attributes to include in it: year and decade.
}
\details{
The \code{include_all} and \code{exclude_all} parameters allow us to include or exclude
all attributes, and then specifically exclude or include the ones we need.
}
\examples{

td <- when() |>
  configure_year_level(decade = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_date_levels}()},
\code{\link{configure_day_level}()},
\code{\link{configure_dimension}()},
\code{\link{configure_month_level}()},
\code{\link{configure_quarter_level}()},
\code{\link{configure_semester_level}()},
\code{\link{configure_week_level}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_time_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
