% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_days}
\alias{date_days}
\title{Precalculated date set}
\format{
A vector.
}
\usage{
date_days
}
\description{
Precalculated date set
}
\seealso{
Other time definition: 
\code{\link{time_seconds}}
}
\concept{time definition}
\keyword{datasets}
