% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{define_day_part}
\alias{define_day_part}
\alias{define_day_part.when}
\title{Define day part}
\usage{
define_day_part(td, hour, name)

\method{define_day_part}{when}(td, hour = NULL, name = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{hour}{A number.}

\item{name}{a string.}
}
\value{
A \code{when} object.
}
\description{
Define day part.
}
\examples{

td <- when() |>
  define_day_part(hour = c(21:23, 0:4), name = "Night")

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{time_seconds}},
\code{\link{when}()}
}
\concept{time definition}
