% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{define_instances}
\alias{define_instances}
\alias{define_instances.when}
\title{Define instances}
\usage{
define_instances(td, start, end, values)

\method{define_instances}{when}(td, start = NULL, end = NULL, values = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{start}{A string, start of the period to be included in the dimension.}

\item{end}{A string, end of the period to be included in the dimension.}

\item{values}{A vector of string.}
}
\value{
A \code{when} object.
}
\description{
Using this function we can define the instances from which the dimension will
be generated according to the rest of its defined characteristics.
}
\details{
We must indicate dates or date components in ISO 8601 format (yyyy-mm-dd). The
times in hh:mm:ss format.
}
\examples{

td_1 <- when() |>
  define_instances(start = "2020", end = "2030")

td_1 <- when() |>
  define_instances(start = "2020-01-01", end = "2030-01-01")

td_2 <- when(type = 'time') |>
  define_instances(values = 1:5)

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
