% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_table}
\alias{generate_table}
\alias{generate_table.when}
\title{Generate table}
\usage{
generate_table(td)

\method{generate_table}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A \code{when} object.
}
\description{
Generate the table of the defined dimension.
}
\examples{

table <- when() |>
  generate_table()

}
\concept{obtaining results}
