% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_data}
\alias{get_data}
\title{Get the data from values and fields}
\usage{
get_data(td, values, fields)
}
\arguments{
\item{td}{A \code{when} object.}

\item{values}{A vector of strings.}

\item{fields}{A vector of strings.}
}
\value{
A \code{tibble}.
}
\description{
Get the data from values and fields
}
\keyword{internal}
