% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_definition.R
\name{get_data_month}
\alias{get_data_month}
\title{Get the data from values and fields}
\usage{
get_data_month(data, values, ...)
}
\arguments{
\item{data}{A \code{tibble}.}

\item{values}{A vector of strings.}

\item{...}{Rest of configuration parameters.}
}
\value{
A \code{tibble}.
}
\description{
Get the data from values and fields
}
\keyword{internal}
