% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_fields}
\alias{get_fields}
\title{Get the fields according to the defined configuration}
\usage{
get_fields(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A vector of string.
}
\description{
Get the fields according to the defined configuration
}
\keyword{internal}
