% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_data.R
\name{get_level_data}
\alias{get_level_data}
\alias{get_level_data.when}
\title{Get the table of the dimension}
\usage{
get_level_data(td)

\method{get_level_data}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A \code{tibble}, the table.
}
\description{
Obtain the table of the defined dimension.
}
\examples{

table <- when() |>
  get_level_data()

}
\concept{obtaining results}
