% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_level_names}
\alias{get_level_names}
\alias{get_level_names.when}
\title{Get level names}
\usage{
get_level_names(td, selected)

\method{get_level_names}{when}(td, selected = FALSE)
}
\arguments{
\item{td}{A \code{when} object.}

\item{selected}{A boolean.}
}
\value{
A string vector.
}
\description{
Returns the names of the levels. We can obtain all the available ones or only
the selected ones.
}
\examples{

names <- when() |>
  get_level_names()

}
\seealso{
Other obtaining results: 
\code{\link{generate_table}()},
\code{\link{get_level_attribute_names}()},
\code{\link{get_table_attribute_names}()},
\code{\link{set_table_attribute_names}()}
}
\concept{obtaining results}
