% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_nexus}
\alias{get_nexus}
\title{Get nexus}
\usage{
get_nexus(name)
}
\arguments{
\item{name}{A string.}
}
\value{
A string.
}
\description{
Given a name, if it ends in "/" the nexus is the empty string, otherwise it
is "/".
}
\keyword{internal}
