% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table}
\alias{get_table}
\alias{get_table.when}
\title{Get the table of the dimension}
\usage{
get_table(td)

\method{get_table}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A \code{tibble}, the table.
}
\description{
Obtain the table of the defined dimension.
}
\examples{

table <- when() |>
  generate_table() |>
  get_table()

}
\seealso{
\code{\link{when}}, \code{\link{generate_table}}

Other exporting results: 
\code{\link{get_table_csv}()},
\code{\link{get_table_rdb}()},
\code{\link{get_table_xlsx}()},
\code{\link{set_attribute_names}()}
}
\concept{exporting results}
