% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_table_attribute_names}
\alias{get_table_attribute_names}
\alias{get_table_attribute_names.when}
\title{Get table attribute names}
\usage{
get_table_attribute_names(td)

\method{get_table_attribute_names}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A string.
}
\description{
Returns the names of the dimension table attributes in string form, so we can
easily use it to rename them if deemed necessary.
}
\examples{

names <- when() |>
  get_table_attribute_names()

}
\seealso{
Other obtaining results: 
\code{\link{generate_table}()},
\code{\link{set_table_attribute_names}()}
}
\concept{obtaining results}
