% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table_csv}
\alias{get_table_csv}
\alias{get_table_csv.when}
\title{Store the table in a csv files}
\usage{
get_table_csv(td, file, type)

\method{get_table_csv}{when}(td, file = NULL, type = 1)
}
\arguments{
\item{td}{A \code{when} object.}

\item{file}{A string, name of a file.}

\item{type}{An integer, 1: uses "." for the decimal point and a comma for the
separator; 2: uses a comma for the decimal point and a semicolon for the
separator.}
}
\value{
A string, name of a file.
}
\description{
Store the table in a csv files
}
\examples{

file <- when() |>
  generate_table() |>
  get_table_csv()

}
\seealso{
Other exporting results: 
\code{\link{get_table_rdb}()},
\code{\link{get_table_xlsx}()},
\code{\link{get_table}()},
\code{\link{set_attribute_names}()}
}
\concept{exporting results}
