% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_definition.R
\name{get_table_day_num_name}
\alias{get_table_day_num_name}
\title{Get the table from values and fields}
\usage{
get_table_day_num_name(table, values, ...)
}
\arguments{
\item{table}{A \code{tibble}.}

\item{values}{A vector of strings.}

\item{...}{Rest of configuration parameters.}
}
\value{
A \code{tibble}.
}
\description{
Get the table from values and fields
}
\keyword{internal}
