% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table_xlsx}
\alias{get_table_xlsx}
\alias{get_table_xlsx.when}
\title{Store the table in a xlsx file}
\usage{
get_table_xlsx(td, file)

\method{get_table_xlsx}{when}(td, file = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{file}{A string, name of a file.}
}
\value{
A string, name of a file.
}
\description{
Store the table in a xlsx file
}
\examples{

file <- when() |>
  generate_table() |>
  get_table_xlsx()

}
\seealso{
Other exporting results: 
\code{\link{get_table_csv}()},
\code{\link{get_table_rdb}()},
\code{\link{get_table}()},
\code{\link{set_attribute_names}()}
}
\concept{exporting results}
