% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_values}
\alias{get_values}
\title{Get the values according to the defined configuration}
\usage{
get_values(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A vector of string.
}
\description{
Get the values according to the defined configuration
}
\keyword{internal}
