% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{get_week_date_range}
\alias{get_week_date_range}
\title{Get week date range}
\usage{
get_week_date_range(start = NULL, end = NULL, type = NULL)
}
\arguments{
\item{start}{A string, start of the period to be included in the dimension.}

\item{end}{A string, end of the period to be included in the dimension.}

\item{type}{A string, type of calendar (NULL, 'iso', 'epi' or 'time').}
}
\value{
A \code{tibble}.
}
\description{
For weeks between the given dates, gets the date of the first and last day of
each week.
}
\examples{

t <- get_week_date_range(start = "2024-01-01", end = "2029-12-31")

}
\seealso{
Other dimension definition: 
\code{\link{define_characteristics}()},
\code{\link{define_instances}()},
\code{\link{get_attribute_definition_function}()},
\code{\link{get_day_part}()},
\code{\link{select_date_levels}()},
\code{\link{select_day_level}()},
\code{\link{select_month_level}()},
\code{\link{select_quarter_level}()},
\code{\link{select_semester_level}()},
\code{\link{select_time_level}()},
\code{\link{select_week_level}()},
\code{\link{select_year_level}()},
\code{\link{set_attribute_definition_function}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
