% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{include_date_level_common}
\alias{include_date_level_common}
\title{Include date level (common)}
\usage{
include_date_level_common(
  td,
  name = NULL,
  level = TRUE,
  include_all = FALSE,
  ...
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{level}{A boolean, include year level.}

\item{include_all}{A boolean, include all fields of the level.}

\item{...}{Rest of boolean configuration parameters.}
}
\value{
A \code{when} object.
}
\description{
Include date level (common)
}
\keyword{internal}
