% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{include_date_levels}
\alias{include_date_levels}
\alias{include_date_levels.when}
\title{Include date levels}
\usage{
include_date_levels(td, day_level, week_level, month_level, year_level)

\method{include_date_levels}{when}(
  td,
  day_level = TRUE,
  week_level = TRUE,
  month_level = TRUE,
  year_level = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{day_level}{A boolean, include day level.}

\item{week_level}{A boolean, include week level.}

\item{month_level}{A boolean, include month level.}

\item{year_level}{A boolean, include year level.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as date type, using this function we can select
the levels to include in it: day, week, month and year.
}
\examples{

td <- when() |>
  include_date_levels(week_level = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
