% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{include_time_level}
\alias{include_time_level}
\alias{include_time_level.when}
\title{Include time}
\usage{
include_time_level(
  td,
  include_time,
  include_minute,
  include_second,
  include_day_part
)

\method{include_time_level}{when}(
  td,
  include_time = TRUE,
  include_minute = TRUE,
  include_second = TRUE,
  include_day_part = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{include_time}{A boolean, include a field for the time.}

\item{include_minute}{A boolean, include the minute level of detail.}

\item{include_second}{A boolean, include the second level of detail.}

\item{include_day_part}{A boolean, include the parts of the day.}
}
\value{
A \code{when} object.
}
\description{
Include time.
}
\examples{

td <- when() |>
  include_time_level()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{time_seconds}}
}
\concept{time definition}
