% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{include_year_level}
\alias{include_year_level}
\alias{include_year_level.when}
\title{Include year level}
\usage{
include_year_level(td, level, include_all, year, decade)

\method{include_year_level}{when}(
  td,
  level = TRUE,
  include_all = FALSE,
  year = TRUE,
  decade = FALSE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{level}{A boolean, include year level.}

\item{include_all}{A boolean, include all fields of the level.}

\item{year}{A boolean, include the year field.}

\item{decade}{A boolean, include the decade field.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a date type, using this function we can select
the year level and its attributes to include in it: year and decade.
}
\examples{

td <- when() |>
  include_year_level(decade = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
