% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{select_date_level_common}
\alias{select_date_level_common}
\title{Configure date level (common)}
\usage{
select_date_level_common(
  td,
  name = NULL,
  include_all = FALSE,
  exclude_all = FALSE,
  ...
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{include_all}{A boolean, include all fields of the level.}

\item{exclude_all}{A boolean, exclude all fields of the level.}

\item{...}{Rest of boolean configuration parameters.}
}
\value{
A \code{when} object.
}
\description{
Configure date level (common)
}
\keyword{internal}
