% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{set_attribute_names}
\alias{set_attribute_names}
\alias{set_attribute_names.when}
\title{Set attribute names}
\usage{
set_attribute_names(td, names)

\method{set_attribute_names}{when}(td, names = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{names}{A string vector.}
}
\value{
A \code{when} object.
}
\description{
Set attribute names
}
\examples{

table <- when() |>
  set_attribute_names()

}
\seealso{
Other exporting results: 
\code{\link{get_table_csv}()},
\code{\link{get_table_rdb}()},
\code{\link{get_table_xlsx}()},
\code{\link{get_table}()}
}
\concept{exporting results}
