% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_seconds}
\alias{time_seconds}
\title{Time in seconds of a day}
\format{
A vector.
}
\usage{
time_seconds
}
\description{
Time in seconds of a day
}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()}
}
\concept{time definition}
\keyword{datasets}
