% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{validate_start_end}
\alias{validate_start_end}
\title{Validate start and end parameters}
\usage{
validate_start_end(td, start, end)
}
\arguments{
\item{td}{A \code{when} object.}

\item{start}{A string, start of the period to be included in the dimension.}

\item{end}{A string, end of the period to be included in the dimension.}
}
\value{
A \code{when} object.
}
\description{
Validate start and end parameters
}
\keyword{internal}
