% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{validate_type}
\alias{validate_type}
\title{Validate type parameter}
\usage{
validate_type(td, type)
}
\arguments{
\item{td}{A \code{when} object.}

\item{type}{A string, type of calendar (NULL, 'iso', 'epi' or 'time').}
}
\value{
A \code{when} object.
}
\description{
Validate type parameter
}
\keyword{internal}
