% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{validate_values}
\alias{validate_values}
\title{Validate values parameter}
\usage{
validate_values(td, values)
}
\arguments{
\item{td}{A \code{when} object.}

\item{values}{A vector of string.}
}
\value{
A \code{when} object.
}
\description{
Validate values parameter
}
\keyword{internal}
