% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when}
\alias{when}
\title{\code{when} S3 class}
\usage{
when(
  name = NULL,
  type = NULL,
  locale = Sys.getlocale("LC_TIME"),
  start = lubridate::today(),
  end = lubridate::today(),
  values = NULL,
  ...
)
}
\arguments{
\item{name}{A string, table name.}

\item{type}{A string, type of calendar (NULL, 'iso', 'epi' or 'time').}

\item{locale}{A locale, to use for day and month names.}

\item{start}{A string, start of the period to be included in the dimension.}

\item{end}{A string, end of the period to be included in the dimension.}

\item{values}{A vector of string.}

\item{...}{Rest of boolean configuration parameters.}
}
\value{
A \code{when} object.
}
\description{
Creates a \code{when} object.
}
\examples{

td <- when()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_day_part}()},
\code{\link{define_instances}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{time_seconds}}
}
\concept{time definition}
