% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{define_instances}
\alias{define_instances}
\alias{define_instances.when}
\title{Define instances}
\usage{
define_instances(td, start, end, values)

\method{define_instances}{when}(td, start = NULL, end = NULL, values = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{start}{A string, start of the period to be included in the dimension.}

\item{end}{A string, end of the period to be included in the dimension.}

\item{values}{A vector of string.}
}
\value{
A \code{when} object.
}
\description{
Define instances.
}
\examples{

td <- when() |>
  define_instances()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{time_seconds}}
}
\concept{time definition}
