% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_table}
\alias{generate_table}
\alias{generate_table.when}
\title{Generate table}
\usage{
generate_table(td)

\method{generate_table}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A \code{when} object.
}
\description{
Once all the characteristics of the dimension have been defined, we can generate
its table according to them using this function.
}
\examples{

td <- when() |>
  generate_table()

}
\seealso{
\code{\link{when}}, \code{\link{get_table}}

Other obtaining results: 
\code{\link{get_level_attribute_names}()},
\code{\link{get_level_names}()},
\code{\link{get_table_attribute_names}()},
\code{\link{set_table_attribute_names}()}
}
\concept{obtaining results}
