% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{get_day_part}
\alias{get_day_part}
\alias{get_day_part.when}
\title{Get day part}
\usage{
get_day_part(td)

\method{get_day_part}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A named vector.
}
\description{
Get day part.
}
\examples{

dp <- when() |>
  get_day_part()

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
