% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{get_day_part}
\alias{get_day_part}
\alias{get_day_part.when}
\title{Get day part}
\usage{
get_day_part(td)

\method{get_day_part}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A named vector.
}
\description{
Get day part.
}
\examples{

dp <- when() |>
  get_day_part()

}
\seealso{
Other dimension definition: 
\code{\link{define_characteristics}()},
\code{\link{define_instances}()},
\code{\link{get_attribute_definition_function}()},
\code{\link{get_week_date_range}()},
\code{\link{select_date_levels}()},
\code{\link{select_day_level}()},
\code{\link{select_month_level}()},
\code{\link{select_quarter_level}()},
\code{\link{select_semester_level}()},
\code{\link{select_time_level}()},
\code{\link{select_week_level}()},
\code{\link{select_year_level}()},
\code{\link{set_attribute_definition_function}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
