% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table}
\alias{get_table}
\alias{get_table.when}
\title{Get the table of the dimension}
\usage{
get_table(td)

\method{get_table}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A \code{tibble}, the table.
}
\description{
Obtain the table of the defined dimension.
}
\examples{

table <- when() |>
  get_table()

}
\seealso{
Other obtaining results: 
\code{\link{generate_table}()}
}
\concept{obtaining results}
