% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table}
\alias{get_table}
\alias{get_table.when}
\title{Get the table of the dimension}
\usage{
get_table(td)

\method{get_table}{when}(td)
}
\arguments{
\item{td}{A \code{when} object.}
}
\value{
A \code{tibble}, the table.
}
\description{
Once all the configuration elements have been defined and the dimension table
has been generated, using this function we can obtain it in \code{tibble} format.
}
\examples{

table <- when() |>
  generate_table() |>
  get_table()

}
\seealso{
\code{\link{when}}, \code{\link{generate_table}}

Other getting results: 
\code{\link{get_table_csv}()},
\code{\link{get_table_rdb}()},
\code{\link{get_table_xlsx}()}
}
\concept{getting results}
