% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_table_attribute_names}
\alias{get_table_attribute_names}
\alias{get_table_attribute_names.when}
\title{Get table attribute names}
\usage{
get_table_attribute_names(td, as_string)

\method{get_table_attribute_names}{when}(td, as_string = TRUE)
}
\arguments{
\item{td}{A \code{when} object.}

\item{as_string}{A boolean.}
}
\value{
A string.
}
\description{
Returns the names of the dimension table attributes as a string vector or in
string form, so we can easily use it to rename them if deemed necessary.
}
\details{
If the table has not been generated yet, returns the attributes it will contain
when it is generated.
}
\examples{

names <- when() |>
  get_table_attribute_names()

}
\seealso{
Other obtaining results: 
\code{\link{generate_table}()},
\code{\link{get_level_attribute_names}()},
\code{\link{get_level_names}()},
\code{\link{set_table_attribute_names}()}
}
\concept{obtaining results}
