% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table_csv}
\alias{get_table_csv}
\alias{get_table_csv.when}
\title{Store the table in a csv files}
\usage{
get_table_csv(td, dir, type)

\method{get_table_csv}{when}(td, dir = NULL, type = 1)
}
\arguments{
\item{td}{A \code{when} object.}

\item{dir}{A string, name of a dir.}

\item{type}{An integer, 1: uses "." for the decimal point and a comma for the
separator; 2: uses a comma for the decimal point and a semicolon for the
separator.}
}
\value{
A string, name of a file.
}
\description{
Once all the configuration elements have been defined and the dimension table
has been generated, using this function we can obtain it in csv format.
}
\details{
If no dir name is given, stores the table in a temporary one.
}
\examples{

file <- when() |>
  generate_table() |>
  get_table_csv()

}
\seealso{
Other getting results: 
\code{\link{get_table_rdb}()},
\code{\link{get_table_xlsx}()},
\code{\link{get_table}()}
}
\concept{getting results}
