% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_table_xlsx}
\alias{get_table_xlsx}
\alias{get_table_xlsx.when}
\title{Store the table in a xlsx file}
\usage{
get_table_xlsx(td, dir)

\method{get_table_xlsx}{when}(td, dir = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{dir}{A string, name of a dir.}
}
\value{
A string, name of a file.
}
\description{
Once all the configuration elements have been defined and the dimension table
has been generated, using this function we can obtain it in \emph{xlsx} format.
}
\details{
If no dir name is given, stores the table in a temporary one.
}
\examples{
\donttest{
file <- when() |>
  generate_table() |>
  get_table_xlsx()
}
}
\seealso{
Other getting results: 
\code{\link{get_table_csv}()},
\code{\link{get_table_rdb}()},
\code{\link{get_table}()}
}
\concept{getting results}
