% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_level.R
\name{include_day_level}
\alias{include_day_level}
\alias{include_day_level.when}
\title{Include day level}
\usage{
include_day_level(
  td,
  day_level,
  include_date,
  include_month_day,
  include_week_day,
  include_day_name,
  include_day_abbr,
  include_day_num_name,
  include_day_num_abbr,
  include_quarter_day,
  include_year_day
)

\method{include_day_level}{when}(
  td,
  day_level = TRUE,
  include_date = TRUE,
  include_month_day = TRUE,
  include_week_day = TRUE,
  include_day_name = TRUE,
  include_day_abbr = TRUE,
  include_day_num_name = TRUE,
  include_day_num_abbr = TRUE,
  include_quarter_day = TRUE,
  include_year_day = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{day_level}{A boolean, include day level.}

\item{include_date}{A boolean, include the date.}

\item{include_month_day}{A boolean, include the day number in the month.}

\item{include_week_day}{A boolean, the day number in the week.}

\item{include_day_name}{A boolean, include the name of the day of the week.}

\item{include_day_abbr}{A boolean, include the name of the day of the week in
abbreviated version.}

\item{include_day_num_name}{A boolean, include the number and name of the day
of the week.}

\item{include_day_num_abbr}{A boolean, include the number and name of the day
of the week in abbreviated version.}

\item{include_quarter_day}{A boolean, include the number of the day in the quarter.}

\item{include_year_day}{A boolean, include the number of the day in the year.}
}
\value{
A \code{when} object.
}
\description{
Include day level.
}
\examples{

td <- when() |>
  include_day_level()

}
\seealso{
Other time definition: 
\code{\link{configure_instances}()},
\code{\link{configure_type}()},
\code{\link{include_date_levels}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{time_seconds}},
\code{\link{when}()}
}
\concept{time definition}
