% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{include_month_level}
\alias{include_month_level}
\alias{include_month_level.when}
\title{Include month level}
\usage{
include_month_level(
  td,
  level,
  include_all,
  month,
  year_month,
  month_name,
  month_num_name,
  month_abbr,
  month_num_abbr,
  quarter,
  year_quarter,
  semester,
  year_semester
)

\method{include_month_level}{when}(
  td,
  level = TRUE,
  include_all = FALSE,
  month = TRUE,
  year_month = TRUE,
  month_name = TRUE,
  month_num_name = TRUE,
  month_abbr = FALSE,
  month_num_abbr = FALSE,
  quarter = FALSE,
  year_quarter = FALSE,
  semester = FALSE,
  year_semester = FALSE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{level}{A boolean, include month level.}

\item{include_all}{A boolean, include all fields of the level.}

\item{month}{A boolean, include the month number.}

\item{year_month}{A boolean, include the year-month combination.}

\item{month_name}{A boolean, include the month name.}

\item{month_num_name}{A boolean, include the month number and name.}

\item{month_abbr}{A boolean, include the month name abbreviated version.}

\item{month_num_abbr}{A boolean, include the month number and name
abbreviated version.}

\item{quarter}{A boolean, include the quarter field.}

\item{year_quarter}{A boolean, include the quarter field.}

\item{semester}{A boolean, include the semester field.}

\item{year_semester}{A boolean, include the semester field.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a date type, using this function we can select
the month level and its attributes to include in it: month, quarter and semester.
}
\details{
We can also obtain the combination of the year with the mentioned attributes.

For the month we have the month number in the year, its name and the abbreviation
of the name. So that the order of the names corresponds to the alphabetical order,
the combination of month number and name and/or abbreviation is included.
}
\examples{

td <- when() |>
  include_month_level(month_abbr = FALSE,
                      month_num_abbr = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
