% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_month.R
\name{include_month_level}
\alias{include_month_level}
\alias{include_month_level.when}
\title{Include month level}
\usage{
include_month_level(
  td,
  month_level,
  include_month,
  include_year_month,
  include_month_name,
  include_month_abbr,
  include_month_num_name,
  include_month_num_abbr,
  include_quarter,
  include_year_quarter,
  include_semester,
  include_year_semester
)

\method{include_month_level}{when}(
  td,
  month_level = TRUE,
  include_month = TRUE,
  include_year_month = TRUE,
  include_month_name = TRUE,
  include_month_abbr = FALSE,
  include_month_num_name = TRUE,
  include_month_num_abbr = FALSE,
  include_quarter = FALSE,
  include_year_quarter = FALSE,
  include_semester = FALSE,
  include_year_semester = FALSE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{month_level}{A boolean, include month level.}

\item{include_month}{A boolean, include the month number.}

\item{include_year_month}{A boolean, include the year-month combination.}

\item{include_month_name}{A boolean, include the month name.}

\item{include_month_abbr}{A boolean, include the month name abbreviated version.}

\item{include_month_num_name}{A boolean, include the month number and name.}

\item{include_month_num_abbr}{A boolean, include the month number and name
abbreviated version.}

\item{include_quarter}{A boolean, include the quarter field.}

\item{include_year_quarter}{A boolean, include the quarter field.}

\item{include_semester}{A boolean, include the semester field.}

\item{include_year_semester}{A boolean, include the semester field.}
}
\value{
A \code{when} object.
}
\description{
Include month level.
}
\examples{

td <- when() |>
  include_month_level()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{include_day_level}()},
\code{\link{time_seconds}}
}
\concept{time definition}
