% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate_key.R
\name{include_surrogate_key}
\alias{include_surrogate_key}
\alias{include_surrogate_key.when}
\title{Include surrogate_key}
\usage{
include_surrogate_key(td, surrogate_key)

\method{include_surrogate_key}{when}(td, surrogate_key = TRUE)
}
\arguments{
\item{td}{A \code{when} object.}

\item{surrogate_key}{A boolean, include a surrogate key in the dimension table.}
}
\value{
A \code{when} object.
}
\description{
Include a generated key in the dimension table. The default name is 'id'. Later
we can rename all the fields in the dimension table.
}
\examples{

td <- when() |>
  include_surrogate_key()

}
\seealso{
Other dimension definition: 
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
