% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate_key.R
\name{include_surrogate_key}
\alias{include_surrogate_key}
\alias{include_surrogate_key.when}
\title{Include surrogate_key}
\usage{
include_surrogate_key(td, surrogate_key)

\method{include_surrogate_key}{when}(td, surrogate_key = TRUE)
}
\arguments{
\item{td}{A \code{when} object.}

\item{surrogate_key}{A boolean, include a surrogate key in the table.}
}
\value{
A \code{when} object.
}
\description{
Include surrogate_key.
}
\examples{

td <- when() |>
  include_surrogate_key()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{time_seconds}},
\code{\link{when}()}
}
\concept{time definition}
