% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{include_time_level}
\alias{include_time_level}
\alias{include_time_level.when}
\title{Include time level}
\usage{
include_time_level(td, time, minute, second, day_part)

\method{include_time_level}{when}(
  td,
  time = TRUE,
  minute = TRUE,
  second = TRUE,
  day_part = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{time}{A boolean, include a field for the time.}

\item{minute}{A boolean, include the minute level of detail.}

\item{second}{A boolean, include the second level of detail.}

\item{day_part}{A boolean, include the parts of the day.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a time type, using this function we can select
its attributes to include in it: time, minute, second and day_part.
}
\details{
The 'hour' attribute will always be included. If the 'minute' attribute is not
included the 'second' attribute will not be included either.
}
\examples{

td <- when() |>
  include_time_level(day_part = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
