% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_week.R
\name{include_week_level}
\alias{include_week_level}
\alias{include_week_level.when}
\title{Include week level}
\usage{
include_week_level(
  td,
  week_level,
  include_week,
  include_year_week,
  include_week_date
)

\method{include_week_level}{when}(
  td,
  week_level = TRUE,
  include_week = TRUE,
  include_year_week = TRUE,
  include_week_date = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{week_level}{A boolean, include week level.}

\item{include_week}{A boolean, include the week number.}

\item{include_year_week}{A boolean, include the year-week combination.}

\item{include_week_date}{A boolean, include the date from which the week is
obtained.}
}
\value{
A \code{when} object.
}
\description{
Include week level.
}
\examples{

td <- when() |>
  include_week_level()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_time_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{time_seconds}}
}
\concept{time definition}
