% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_date.R
\name{include_week_level}
\alias{include_week_level}
\alias{include_week_level.when}
\title{Include week level}
\usage{
include_week_level(td, level, include_all, week, year_week, week_date)

\method{include_week_level}{when}(
  td,
  level = TRUE,
  include_all = FALSE,
  week = TRUE,
  year_week = TRUE,
  week_date = FALSE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{level}{A boolean, include week level.}

\item{include_all}{A boolean, include all fields of the level.}

\item{week}{A boolean, include the week number.}

\item{year_week}{A boolean, include the year-week combination.}

\item{week_date}{A boolean, include the date from which the week is
obtained.}
}
\value{
A \code{when} object.
}
\description{
When the dimension is defined as a date type, using this function we can select
the week level and its attributes to include in it: week, year_week and week_date.
}
\details{
For the first and last days of the year, the year associated with the week may
be different from the year of the date, depending on the date type selected.

The week number associated with each date depends on the type of date dimension
selected: standard (NULL), ISO 8601 ('iso') or epidemiological ('epi').

The standard week numbers blocks of 7 days beginning on January 1. The last week
of the year can be less than 7 days long.

The ISO 8601 week numbers blocks of 7 days from Monday to Sunday. The first and
last week of the year can contain days from the previous or next year.

The epidemiological week is like ISO 8601 only that it considers that the week
begins on Sunday, it numbers blocks from Sunday to Monday.
}
\examples{

td <- when() |>
  include_week_level(week_date = FALSE)

}
\seealso{
Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_time_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()},
\code{\link{when}()}
}
\concept{dimension definition}
