% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_year.R
\name{include_year_level}
\alias{include_year_level}
\alias{include_year_level.when}
\title{Include year level}
\usage{
include_year_level(td, year_level, include_year, include_decade)

\method{include_year_level}{when}(
  td,
  year_level = TRUE,
  include_year = TRUE,
  include_decade = TRUE
)
}
\arguments{
\item{td}{A \code{when} object.}

\item{year_level}{A boolean, include year level.}

\item{include_year}{A boolean, include the year field.}

\item{include_decade}{A boolean, include the decade field.}
}
\value{
A \code{when} object.
}
\description{
Include year level.
}
\examples{

td <- when() |>
  include_year_level()

}
\seealso{
Other time definition: 
\code{\link{configure_dimension}()},
\code{\link{define_day_part}()},
\code{\link{define_instances}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{time_seconds}},
\code{\link{when}()}
}
\concept{time definition}
