% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{set_day_part}
\alias{set_day_part}
\alias{set_day_part.when}
\title{Set day part}
\usage{
set_day_part(td, hour, name)

\method{set_day_part}{when}(td, hour = NULL, name = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{hour}{A number, hour number (between 0 and 23).}

\item{name}{a string, name of the part of the day.}
}
\value{
A \code{when} object.
}
\description{
Using this function we can change the name assigned to the hours of the day to
designate the parts of the day.
}
\examples{

td <- when() |>
  set_day_part(hour = c(21:23, 0:4), name = "Night")

}
\seealso{
Other dimension definition: 
\code{\link{define_characteristics}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{select_date_levels}()},
\code{\link{select_day_level}()},
\code{\link{select_month_level}()},
\code{\link{select_quarter_level}()},
\code{\link{select_semester_level}()},
\code{\link{select_time_level}()},
\code{\link{select_week_level}()},
\code{\link{select_year_level}()},
\code{\link{when}()}
}
\concept{dimension definition}
