% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_time.R
\name{set_day_part}
\alias{set_day_part}
\alias{set_day_part.when}
\title{Set day part}
\usage{
set_day_part(td, hour, name)

\method{set_day_part}{when}(td, hour = NULL, name = NULL)
}
\arguments{
\item{td}{A \code{when} object.}

\item{hour}{A number.}

\item{name}{a string.}
}
\value{
A \code{when} object.
}
\description{
Set day part.
}
\examples{

td <- when() |>
  set_day_part(hour = c(21:23, 0:4), name = "Night")

}
\seealso{
Other dimension definition: 
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_surrogate_key}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{when}()}
}
\concept{dimension definition}
