% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when}
\alias{when}
\title{\code{when} S3 class}
\usage{
when(
  name = NULL,
  type = NULL,
  locale = Sys.getlocale("LC_TIME"),
  start = lubridate::today(),
  end = lubridate::today(),
  values = NULL,
  ...
)
}
\arguments{
\item{name}{A string, table name.}

\item{type}{A string, type of calendar (NULL, 'iso', 'epi' or 'time').}

\item{locale}{A locale, to use for day and month names.}

\item{start}{A string, start of the period to be included in the dimension.}

\item{end}{A string, end of the period to be included in the dimension.}

\item{values}{A vector of string.}

\item{...}{Rest of boolean configuration parameters.}
}
\value{
A \code{when} object.
}
\description{
Creates a \code{when} object.
}
\details{
Using the parameters of this function we can configure practically  all the
elements of the dimension. Alternatively, we can use the configuration functions
to define the available options.

We discuss the parameters in each of the specific configuration functions.
}
\examples{

td_1 <- when()

td_2 <- when(type = 'time')

}
\seealso{
\code{\link{generate_table}}, \code{\link{get_table}}

Other dimension definition: 
\code{\link{configure_dimension}()},
\code{\link{define_instances}()},
\code{\link{get_day_part}()},
\code{\link{include_date_levels}()},
\code{\link{include_day_level}()},
\code{\link{include_month_level}()},
\code{\link{include_time_level}()},
\code{\link{include_week_level}()},
\code{\link{include_year_level}()},
\code{\link{set_day_part}()}
}
\concept{dimension definition}
